SHA format invalid on Windows
When calling `git log`, the format was set to `'%h'` which printed the SHA surrounded by single quotes, e.g. `'c665097'`. The regex would pick this up as the actual SHA, and try to pass it to `git show`. This works fine on OS X (I guess the shell is handling the quotes), but fails on Windows with an invalid format error. This change removes the single quotes. Tested on Windows and OS X.