Replace parsing of `git` commands output with ruby-git library
Instead of parsing output of `git` commands, which are liable to change
over time, we want to instead use a library, in this case `ruby-git`,
which gives us a more straightforward interface.

Use an instance variable to ensure we're only using a shared object
that's set up and read once.

Pull minor bumps on the ruby-git library, as they will be backwards
compatible via SemVer conventions.

Notice that merge commits are treated differently now - we're picking up
the actual merge commit, rather than the parent, as was seen previously.