Add test for merge commits
Merge commits show the parent commit (from the branch) rather than the
actual merge commit itself. However, this isn't covered by any test
cases.

This requires some refactoring to allow the `dot_git_path` to change, as
well as ensuring that we always clean out the `.git` directory in our
`test_repo` so we always have a clean working state, instead of risking
contamination between test runs.

Note that we need the `.gitkeep` file in `refs/heads` as the folder is
empty, but Git needs that folder to exist, otherwise it won't be
recognised as a Git repo.